fixing broken unit tests
